QI.bu6of4i: we now impose a 30-second timeout (more or less) on rendering any given QL expression.
Also QI.7w4gd72: QL now copes with Large Number literals.

+++ querki/scalajvm/app/querki/ql/AST.scala
@@ -16,7 +16,7 @@ import querki.values.QLContext
 // Marker trait for the types that can be results of parsing a complete QLText.
 sealed trait QLParseResultVal
 
-case class QLNumber(n:Int) extends QLStage(None) {
+case class QLNumber(n:Long) extends QLStage(None) {
   def reconstructString = n.toString
 }
